add maven assembly plugin
